Add files via upload
This report documents the analysis of a credit card
fraud dataset and the algorithm’s prediction of fraud by using neural nets and machine learning. This document is a research paper for a final project for the class STP 494/598: Machine Learning and Statistical Analysis taught by Dr. Robert McCulloch. The students in this group had no previous experience in R, Machine Learning, or Neural Nets and used the project as a test of how far they have progressed throughout the semester.